requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv==1.2.1
pydantic==2.12.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
